refactor: move `builder_error_to_backend` to `backend/mod.rs`
Eliminates the asymmetry where `ffi.rs` used a named helper and `native.rs`
inlined an identical three-arm match. Both backends now call
`super::builder_error_to_backend`.

Addresses manki-review comment on PR #173
https://github.com/xdustinface/dash-spv-wallet/pull/173#discussion_r3263304098
